Refactor predefined executors into separate class
